Update README.md
An analytical and performance-driven professional with academic and project experience in data science and mathematics. 
Data Science professional with a master’s degree in mathematics, and a track record in leveraging technical skills in Python, R, SAS and machine learning to manipulate large datasets, generate insights, and improve automation. Demonstrated experience in tech driven roles in the customer experience and business analysis sectors.